Python3, file name change
Although the Jupyter Notebook says to use Python 3, this script would
not run because ConfigParser was renamed to configparser in 3, and
SafeConfigParser was renamed to ConfigParser.  I fixed the imports so
this runs in Python 3 and above now.

Also, there was an issue not with this script, per se, but with the
default file-name created by the collector .html script.  The files
save as raw_DATE_sample.json, but the script looks for
raw_DATA_sample.json.  I added a line to fix the typo in the filename